Muda o tipo do atributo home no perfil para um array de float

+++ src/schema/profile.ts
@@ -39,7 +39,7 @@ export const ProfileTypeDef = `
     gender: Gender
     posts: [Post!]!
     photoURL: String
-    home: [Int]
+    home: [Float]
     score: Int!
     scoreBalance: Int!
   }
